Add curio, formatter-date and formatter-number
This adds curio, a small enumerable collection library. That can be used
as an in-memory dictionary for objects.

This adds formatter-date, a date formatter with support for time zones.

This adds formatter-number, a simple number formatter with the goal to
support all kind of number formatting without any third party
extensions.